Fixed loading of Gothic 2 OU.BIN

+++ zenload/zTypes.h
@@ -157,7 +157,7 @@ namespace ZenLoad
 
 	struct oCMsgConversationData : public ParsedZenObject
 	{
-		uint8_t subType;
+		uint32_t subType;
 		std::string text;
 		std::string name;
 	};
@@ -166,7 +166,7 @@ namespace ZenLoad
 	{
 		bool mustBeAlive;
 		std::string roleName;
-		uint8_t roleType;
+		uint32_t roleType;
 	};
 
 	struct zCCSBlockSyncData : public ParsedZenObject
